requirements.txt
alabaster==0.7.12
astroid==2.11.4
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.10.1
bandit==1.7.4
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cohesion==1.0.0
colorama==0.4.4
coverage==6.3.2
dill==0.3.4
docutils==0.17.1
flake8==4.0.1
flake8-docstrings==1.6.0
flake8-polyfill==1.0.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.0
mando==0.6.4
Markdown==3.3.6
MarkupSafe==2.1.1
mccabe==0.6.1
mock==4.0.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
packaging==21.3
pathspec==0.9.0
pbr==5.8.1
pdoc3==0.10.0
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
pylint==2.13.8
pyparsing==3.0.8
pytest==7.1.2
pytz==2022.1
PyYAML==6.0
radon==5.1.0
requests==2.27.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stevedore==3.5.0
tk==0.1.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
utils==1.0.1
Utils-Config==1.0.0
wrapt==1.14.1
zipp==3.8.0
